requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
pre-commit=4.1.0
